+++ behaviors/wysiwyg.js
@@ -1,10 +1,10 @@
 var _ = require('lodash'),
   select = require('selection-range'),
   MediumEditor = require('medium-editor'),
-  MediumButton = require('../services/medium-button'),
+  MediumButton = require('../services/field-helpers/medium-button'),
   dom = require('@nymag/dom'),
   db = require('../services/edit/db'),
-  render = require('../services/render'),
+  render = require('../services/components/render'),
   focus = require('../decorators/focus'),
   references = require('../services/references'),
   edit = require('../services/edit'),
