+++ client.js
@@ -2,11 +2,11 @@
 // note: use strict above is applied to the whole browserified doc
 var nodeUrl = require('url'),
   references = require('./services/references'),
-  behaviors = require('./services/behaviors'),
-  decorators = require('./services/decorators'),
+  behaviors = require('./services/forms/behaviors'),
+  decorators = require('./services/components/decorators'),
   dom = require('@nymag/dom'),
   EditorToolbar = require('./controllers/kiln-toolbar'),
-  render = require('./services/render'),
+  render = require('./services/components/render'),
   progress = require('./services/progress'),
   Konami = require('konami-js'),
   pane = require('./services/pane'),
