+++ decorators/focus.test.js
@@ -1,7 +1,7 @@
 var dirname = __dirname.split('/').pop(),
   filename = __filename.split('/').pop().split('.').shift(),
   forms = require('../services/forms'),
-  select = require('../services/select'),
+  select = require('../services/components/select'),
   references = require('../services/references'),
   lib = require('./focus');
