+++ services/components/add-component-handler.test.js
