+++ services/components/groups.test.js
@@ -1,7 +1,7 @@
 var dirname = __dirname.split('/').pop(),
   filename = __filename.split('/').pop().split('.').shift(),
   lib = require('./groups'),
-  references = require('./references');
+  references = require('../references');
 
 describe(dirname, function () {
   describe(filename, function () {
