+++ services/components/select.js
@@ -2,16 +2,16 @@
 // components that are "selected" get their bar (and their parents' bars) shown
 
 var _ = require('lodash'),
-  references = require('./references'),
+  references = require('../references'),
   dom = require('@nymag/dom'),
-  tpl = require('./tpl'),
-  edit = require('./edit'),
-  focus = require('../decorators/focus'),
-  placeholder = require('../decorators/placeholder'),
-  forms = require('./forms'),
+  tpl = require('../tpl'),
+  edit = require('../edit'),
+  focus = require('../../decorators/focus'),
+  placeholder = require('../../decorators/placeholder'),
+  forms = require('../forms'),
   groups = require('./groups'),
-  label = require('./label'),
-  scrollToY = require('./scroll').toY,
+  label = require('../label'),
+  scrollToY = require('../scroll').toY,
   addComponentHandler = require('./add-component-handler'),
   hidden = 'kiln-hide',
   currentSelected;
