+++ services/components/select.test.js
@@ -1,12 +1,12 @@
 var dirname = __dirname.split('/').pop(),
   filename = __filename.split('/').pop().split('.').shift(),
-  references = require('./references'),
+  references = require('../references'),
   groups = require('./groups'),
-  edit = require('./edit'),
-  forms = require('./forms'),
-  focus = require('../decorators/focus'),
-  site = require('./site'),
-  tpl = require('./tpl'),
+  edit = require('../edit'),
+  forms = require('../forms'),
+  focus = require('../../decorators/focus'),
+  site = require('../site'),
+  tpl = require('../tpl'),
   dom = require('@nymag/dom'),
   lib = require('./select'),
   hidden = 'kiln-hide';
