+++ services/forms/index.js
@@ -1,11 +1,11 @@
 var _ = require('lodash'),
   dom = require('@nymag/dom'),
-  references = require('./references'),
+  references = require('../references'),
   formCreator = require('./form-creator'),
-  edit = require('./edit'),
-  render = require('./render'),
+  edit = require('../edit'),
+  render = require('../components/render'),
   formValues = require('./form-values'),
-  groups = require('./groups'),
+  groups = require('../components/groups'),
   inlineSelector = '.editor-inline',
   overlaySelector = '.editor-overlay-background',
   currentForm = {}; // used to track if changes have been made.
